requirements.txt
pydantic[email]
fastapi
uvicorn
pymongo
pydantic_settings
passlib
motor
pyjwt